requirements.txt
# Core Dependencies for LLM Extraction
# NOTE: Critical packages (numpy, opencv, paddlepaddle) use binary wheels only
# This prevents long compilation times and ensures compatibility
python-dotenv>=1.0.0
openai>=1.0.0

# PDF and Image Processing
pdf2image>=1.16.0
Pillow>=10.0.0

# OCR Dependencies - CRITICAL: Use exact versions for compatibility
# NumPy MUST be <2.0.0 due to ABI compatibility with OpenCV and PaddlePaddle
numpy==1.26.4

# PaddleOCR and PaddlePaddle - Core OCR functionality
# Note: Installing paddleocr without PyMuPDF dependency (we don't use PDF features)
paddlepaddle==2.6.2
# paddleocr==2.7.0.3  # Commented out - will install manually without PyMuPDF

# OpenCV - MUST match PaddleOCR requirements (<=4.6.0.66)
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66

# PaddleOCR Required Dependencies
attrdict
beautifulsoup4
cython
fire>=0.3.0
imgaug
lxml
openpyxl
pdf2docx
premailer
python-docx
scikit-image
shapely>=2.0.0
pyclipper>=1.3.0
lmdb>=1.4.0
tqdm>=4.66.0
visualdl>=2.5.0
python-Levenshtein>=0.21.0
rapidfuzz>=3.0.0

# Property-Based Testing (optional)
hypothesis>=6.0.0
reportlab>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Database Dependencies
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0

# Streamlit UI Dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
dependency-injector>=4.41.0
python-multipart>=0.0.6
httpx>=0.25.0

# Authentication
authlib>=1.2.0
itsdangerous>=2.1.0

# Data Visualization
plotly>=5.17.0
pandas>=2.0.0

# PDF Generation
fpdf2>=2.7.0

# KNOWN CONFLICTS (non-critical):
# - PyMuPDF: paddleocr requires <1.21.0 which doesn't build on Python 3.12
#   SOLUTION: Install paddleocr with --no-deps (we don't use its PDF features)
#   We use pdf2image for PDF processing instead
# - opencv-python-headless: If installed, will require numpy>=2.0.0
#   DO NOT install opencv-python-headless - use opencv-python instead
 
 